Rework setup, fix shader setup.
Layouts are now their own class. Pick from the switch statement in setup(). Right there's only default and growing. 

Resize glitch is still there but it was being complicated by other problems in the grid logic so it should be easier to tackle now.

Also got rid of the flex box and made the resize code a whole lot better. Should make it easier to start bringing other CSS elements in.